Make single preview view page.
